// src/components/utils/RightTriangleBtn.d.ts
import React from 'react';
declare const RightTriangleBtn: React.FC<{
    className?: string;
    children: React.ReactNode;
    onClick: Function;
    style?: React.CSSProperties;
}>;
export declare const RightTriangleUpBtn: React.FC<{
    className?: string;
    children: React.ReactNode;
    onClick: Function;
    style?: React.CSSProperties;
}>;
export default RightTriangleBtn;