// src/components/utils/LeftTriangleBtn.d.ts
import React from 'react';
declare const LeftTriangleBtn: React.FC<{
    className?: string;
    children: React.ReactNode;
    onClick: Function;
    style?: React.CSSProperties;
}>;
export declare const LeftTriangleUpBtn: React.FC<{
    className?: string;
    children: React.ReactNode;
    onClick: Function;
    style?: React.CSSProperties;
}>;
export default LeftTriangleBtn;